requirements.txt
torch
python-chess
cairosvg
numpy